requirements-test.txt
-r requirements.txt

colorlog
pytest
pytest-homeassistant-custom-component
